third_party/blink/web_tests/external/wpt/css/css-color-adjust/meta/meta-color-scheme-insert-expected.txt

This is a testharness.js-based test.
FAIL Computed root color-scheme should be normal assert_equals: expected (string) "normal" but got (undefined) undefined
FAIL Computed root color-scheme should be light assert_equals: expected (string) "light" but got (undefined) undefined
FAIL Computed root color-scheme should be dark assert_equals: expected (string) "dark" but got (undefined) undefined
Harness: the test ran to completion.
